feat(everything): Refactoring the way events are read to enable multiple event types
