+++ src/commonMain/kotlin/com/monta/slack/notifier/model/serializers/BaseGithubContext.kt
@@ -1,0 +1,9 @@
+package com.monta.slack.notifier.model.serializers
+
+import kotlinx.serialization.Serializable
+
+class BaseGithubContext(
+    val displayName: String?,
+    val sha: String?,
+    val message: String?,
+)
